fix: NPC repeat-dialogue OOB crash + sim dashboard float averages (P1-B)
Four defects found while running the 500-run baseline:
- A: start_dialogue iterated i<5 over the 2-slot repeat_dialogue pool
  (no nullptr terminator) -> wild-pointer strlen -> SIGSEGV. Any real
  player re-talking to an NPC hits this too, not sim-only. Now iterates
  by actual pool size via std::size.
- B: spawn_floor_npcs wrote tile coords into _npc_count-1 (the slot of
  whichever NPC was created last, not this NPC) -> misplaced NPCs.
  Now resolves the slot by npc_id.
- C: new_game() never reset NPC state (_npc_state/_npc_count/_dialogue)
  -> cross-run residue; sim shares one GameScene across 100 runs.
- D: BalanceReport avg_* fields were int (1x2046-dmg run + 99x0 runs
  averaged to 0) and avg_heal was never serialized to JSON.

Verified: 60/60 ctest, world validator 0 errors, same-seed double-run
MD5 identical, all 5 baseline seeds (3/7/11/19/23) complete 100 runs
exit=0 (previously 4/5 seeds crashed within the first few runs).

+++ src/core/sim/sim_runner.h
@@ -128,13 +128,15 @@ struct BalanceReport {
     int total_runs = 0;
     int total_wins = 0;
     float win_rate = 0;
-    int avg_floor = 0;
-    int avg_turns = 0;
-    int avg_damage_dealt = 0;
-    int avg_damage_taken = 0;
-    int avg_heal = 0;
-    int avg_relics = 0;
-    int avg_equipment = 0;   // Q3.2: 平均每局自动装备数
+    // P1-B-fix(D): 均值改 float — 原int增量均值在"少数局有大量/多数局为0"分布下
+    // 每步截断归零 (s7: 1局2046伤害99局0 → avg 应20.46, int算出0), avg_heal 也被截丢
+    float avg_floor = 0;
+    float avg_turns = 0;
+    float avg_damage_dealt = 0;
+    float avg_damage_taken = 0;
+    float avg_heal = 0;
+    float avg_relics = 0;
+    float avg_equipment = 0;   // Q3.2: 平均每局自动装备数
     int boss_kill_count[3] = {0};          // F5, F10, F15 counts
     int death_floor_dist[16] = {0};        // floor 1-15
     int build_count[13] = {0};             // 12 BuildType counts (index 0 unused)
